+++ com.unity.netcode.gameobjects/Tests/Runtime/ConnectionApproval.cs
@@ -34,7 +34,7 @@ namespace Unity.Netcode.RuntimeTests
 
         protected override int NumberOfClients => 1;
 
-        private Guid m_ValidationToken;
+        private string m_ValidationToken;
 
         protected override bool ShouldCheckForSpawnedPlayers()
         {
@@ -46,8 +46,13 @@ namespace Unity.Netcode.RuntimeTests
             m_ClientDisconnectReasonValidated = false;
             m_BypassConnectionTimeout = m_PlayerCreation == PlayerCreation.FailValidation;
             m_Validated.Clear();
-            m_ValidationToken = Guid.NewGuid();
-            var validationToken = Encoding.UTF8.GetBytes(m_ValidationToken.ToString());
+            m_ValidationToken = string.Empty;
+
+            while (m_ValidationToken.Length < 2000)
+            {
+                m_ValidationToken += Guid.NewGuid().ToString();
+            }
+            var validationToken = Encoding.UTF8.GetBytes(m_ValidationToken);
             m_ServerNetworkManager.ConnectionApprovalCallback = NetworkManagerObject_ConnectionApprovalCallback;
             m_ServerNetworkManager.NetworkConfig.PlayerPrefab = m_PlayerCreation == PlayerCreation.Prefab ? m_PlayerPrefab : null;
             if (m_PlayerCreation == PlayerCreation.PrefabHash)